Add COLORREF helper class and move COLORREF methods there
